feat: organizaçao do customtable
