+++ src/components/CustomTable/components/DefaultTable/utils/handleCellChange.ts
@@ -1,0 +1,73 @@
+import Handsontable from "handsontable";
+import { isEquivalent } from "../../../../../utils";
+
+const handleCellChange: any = async (
+  changes: any,
+  hotInstance: Handsontable | null | undefined,
+  isTableLocked: boolean,
+  setIsTableLocked: React.Dispatch<React.SetStateAction<boolean>>,
+  handleSave: (value: any, isNew: boolean, productId: string) => Promise<any>,
+  dataProvider: any[],
+  setDataProvider: React.Dispatch<React.SetStateAction<any[]>>,
+) => {
+  if (changes !== null && changes.length && !isTableLocked && hotInstance) {
+    const isNew = !!dataProvider[changes[0][0]].id;
+    const customChanges = changes as Handsontable.CellChange[];
+    if (
+      typeof customChanges[0][2] === "object" &&
+      typeof customChanges[0][3] === "object" &&
+      !isEquivalent(customChanges[0][2], customChanges[0][3])
+    ) {
+      try {
+        if (!isNew) setIsTableLocked(true);
+        const id = await handleSave(
+          dataProvider[customChanges[0][0]],
+          isNew,
+          dataProvider[customChanges[0][0]]?.id,
+        );
+        if (
+          id &&
+          /^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$/.test(
+            id.toString(),
+          )
+        ) {
+          const updated = dataProvider;
+          updated[customChanges[0][0]].id = id;
+          setDataProvider(updated);
+        }
+      } finally {
+        if (!isNew) setIsTableLocked(false);
+        console.log("BF");
+      }
+    }
+    if (
+      typeof customChanges[0][2] !== "object" &&
+      customChanges[0][2] !== customChanges[0][3] &&
+      dataProvider.length
+    ) {
+      console.log("Changes", { changes });
+      try {
+        if (!isNew) setIsTableLocked(true);
+        const id = await handleSave(
+          dataProvider[customChanges[0][0]],
+          isNew,
+          dataProvider[customChanges[0][0]]?.id,
+        );
+        if (
+          id &&
+          /^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$/.test(
+            id.toString(),
+          )
+        ) {
+          const updated = dataProvider;
+          updated[customChanges[0][0]].id = id;
+          setDataProvider(updated);
+        }
+      } finally {
+        if (!isNew) setIsTableLocked(false);
+      }
+    }
+  }
+};
+
+export default handleCellChange;
