+++ src/components/CustomTable/components/DefaultTable/utils/hasAtLeastOneProduct.ts
@@ -1,0 +1,4 @@
+function hasAtLeastOneProduct(dataProv: any[]): boolean {
+  return dataProv?.length > 0;
+}
+export default hasAtLeastOneProduct;
